Be able to construct an OmiseObject from a given array data.
Previously there is only one way that can be executed in order to construct a new OmiseObject, that's to fetch (retrieve) from Omise API.
Meaning that we cannot fully use the benefit of retrieving a data list (for example, charge list) as we cannot reconstruct an OmiseObject from a given charge data.
For example, we cannot do the following:
$charges = OmiseCharge::retrieve();

foreach($charges as $charge) {
    $charge->capture();
}

As we do not have a way to reconstruct the data from the charge list back to OmiseObject.